[リファクタリング] Json文字列への変換処理をJsonMessageCreaterに任せた
レビューを取り込み変更
https://github.com/kmurayama-aim/senka-hal-tokyo/pull/32#discussion_r245585207
WebSocketInitializerがJson文字列への変換処理を行っていたが、
初期化以外のことをやるべきでは無いため変更
