+++ Client/Assets/Scripts/JsonMessageCreater.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc1e00f4592838f4aaed80ddf8922e03
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
